Move traces to separate package
