+++ tests/test_utils.py
@@ -21,7 +21,7 @@ from perun.fuzz import filetype
 from perun.logic import commands, config
 from perun.testing import asserts
 from perun.utils import log
-from perun.utils.common import common_kit, cli_kit
+from perun.utils.common import common_kit, cli_kit, traces_kit
 from perun.utils.exceptions import (
     SystemTapScriptCompilationException,
     SystemTapStartupException,
@@ -371,11 +371,11 @@ def test_traces():
         {"func": "__mod_lruvec_state"},
         {"func": "__mod_memcg_lruvec_state"},
     ]
-    assert common_kit.compute_distance(trace_a, trace_a) == 0
-    assert common_kit.compute_distance(trace_a, trace_b) == 1.4
-    assert common_kit.compute_distance(trace_b, trace_a) == 1.4
-    assert common_kit.compute_distance(trace_a, trace_c) == 2.0
-    assert common_kit.compute_distance(trace_b, trace_c) == 3.4
+    assert traces_kit.compute_distance(trace_a, trace_a) == 0
+    assert traces_kit.compute_distance(trace_a, trace_b) == 1.4
+    assert traces_kit.compute_distance(trace_b, trace_a) == 1.4
+    assert traces_kit.compute_distance(trace_a, trace_c) == 2.0
+    assert traces_kit.compute_distance(trace_b, trace_c) == 3.4
 
 
 def test_machine_info(monkeypatch):
